@@ -5,11 +5,16 @@ declare(strict_types=1);
 namespace Knp\DictionaryBundle\DataCollector;
 
 use Generator;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\DataCollector\DataCollector;
+use Throwable;
 
 final class DictionaryDataCollector extends DataCollector
 {
-    use DictionaryDataCollector\SymfonyCompatibilityTrait;
+    public function collect(Request $request, Response $response, Throwable $exception = null): void
+    {
+    }
 
     /**
      * @param array<mixed> $keys
